# Shared/Dtos/Report/TextSummaryDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Shared.Dtos.Report
{
    public record TextSummaryDto(
         int TotalTexts,
         int ViolentIncidents,
         int AgainstViolenceAnalyses,
         int NeutralTextAnalyses
     );
}